+++ road-trip-api/src/main/java/road/trip/api/trip/request/TripEditRequest.java
@@ -18,7 +18,7 @@ public class TripEditRequest {
     private AdventureLevel adventureLevel;
 
     @JsonProperty("duration")
-    private Integer driveDuration;
+    private Double driveDuration;
 
     private Double distance;
